+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 3.0.0 (unreleased)
+## 3.0.0 (2017-10-18)
 
 **Enhancements:**